[WTF] card.h: Added a missing #pragma pack().
This is what caused all the weird issues with QImage before!

I encountered the problem again with QString and QByteArray. It turns out
the problem was card.h was included before the Qt headers, so anything that
included card.h ended up packing the Qt structs. Qt doesn't used packed
structs, so that resulted in a massive failure.