GcImage::FromCI8(): Optimized to use BlitTile() instead of QPainter.
Benchmarking was done in the same way as the updated RGB5A3 function.

Time for 10,000 iterations converting a 96x32 CI8 image:
- QPainter version: 1.991825 seconds
- BlitTile version: 0.045806 seconds

The speedup is 43.5x. Part of this is because the color table is only
applied once with the updated version, whereas it was applied with every
tile in the QPainter version.

BlitTile(): Fixed a bug that resulted in an error if W > H. This didn't
show up with RGB5A3, since that format uses 4x4 tiles, but it showed up
with CI8, which uses 8x4 tiles.