GcnMcFileDb: Added a field gcn_file_def::regions.
This is a bitfield indicating what regions the file definition is
valid for. This will be used for games that have identical save
files in e.g. USA and EUR.

RegionCharToBitfield(): Convert a region character, e.g. 'J', 'E',
'P', or 'K', to a region_t bitfield value.

TODO: Add a <regions> element to allow specifying multiple regions.
Currently, the regions field is only populated with the region of
the specified game code.