GcImage::RGB5A3_to_ARGB32(): Correctly support RGB4A3 mode.
If the high bit in an RGB5A3 pixel is not set, the pixel should be
handled as RGB4A3. If it is set, the pixel should be handled as
RGB5 (i.e. how it was handled before, but without transparency).

This fixes the memory card icon for "Action Replay".