SearchThread: Initial async implementation.
It seems that most of the slowdown seen while seraching with
the synchronous implementation was caused by GUI updates.
The asynchronous implementation runs really fast, but the UI
is still a bit slow, since it updates for every block.

This will probably change once I add enough entries to the database.

McRecoverWindow: Move the card_direntry adding code to
searchThread_searchFinished_slot(). This slot is called for both
synchronous and asynchronous modes, so we don't have to code a
special path for either one.

TODO:
- QThread isn't shut down properly.
- Scanning more than once doesn't work.