Added a new algorithm to skip over used blocks.
This should help improve recovery if the memory card has fragmented files.

If the algorithm ends up wrapping around itself due to too many valid
and/or fragmented files, it will fall back to the old algorithm.

Note that the used blocks map is *not* updated if the file wraps around
the end of the card. This is done to prevent collisions with files at
the beginning of the card.

MemCardFile: Replaced the start/length constructor with fatEntries.