McRecoverWindow: Added a QSplitter layout.
This allows the user to resize the split between the card info panel
and the file QTreeView.

Set the initial size of the card info panel to 160px.

Set the stretch factor of the card info panel to 0. This ensures that
only the QTreeView will stretch when the window is resized, and the
card info panel will remain at its current size.