card.h: Added more card_header fields.
byteswap.h: Added 64-bit byteswapping macros.

MemCard: Byteswap the new card_header fields.

NOTE: The formatTime field's units is dependent on the CPU clock
of the system formatting the card, i.e. it will be based on 485 MHz
for GCN or Wii in GCN mode, or 729 MHz for Wii in Wii mode.