GcnMcFileDb: Added support for multiple checksum instances.
If a <checksum> tag has a <multiple> tag, that checksum will
be duplicated for a given number of instances with a specified
increment for each instance.

This is useful for e.g. Super Mario Sunshine and Luigi's Mansion,
which have multiple checksums all with the same length and algorithm
at the same position in each block.