data/GcnMcFileDb.xml: Added a <dbInfo> section.
This section indicates the database format version, as well as
the database name, description, version, and author.

NOTE: This *must* be the first section within GcnMcFileDb.
If it isn't, the file will be rejected. In addition, the
dbFormatVersion major number must match the program, and
the minor number must be less than or equal to the program.