McRecoverWindow: Disable the QMenuBar border hiding for now.
It looks ugly on the classic Windows theme, but the border hiding
makes it look even worse with Windows Aero: the bar is truncated.

I noticed that this happened on Linux with certain themes, which
is why I limited it to Win32. Unfortunately, it happens on Aero
as well, so I'll need to figure out a better way to eliminate the
extra menu bar border on Windows Classic without it screwing up
on other themes.