McRecoverWindow: Promote the Memory Card toolbar to the MainWindow toolbar.
The toolbar will also be used for selecting the primary region,
so we need to move it to the MainWindow in order to have enough
space to show all of the buttons.

NOTE: actionAbout isn't part of the toolbar in the UI file, since
it's added after a spacer.

Also, make sure actionExit uses a themed icon.