GcnMcFileDb: Added initial multiple files support.
GcnMcFileDb will return all possible matches in a QVector<SearchData>.
Note that the SearchData's fatEntries is not initialized, so the
SearchThreadWorker has to make a copy of SearchData and initialize
the fatEntries vector.

Also, SearchThreadWorker currently only checks the first entry,
so it acts exactly as it did before.

TODO:
- Add "preferred region" support.
- Add multiple database file support.