Use Q_DECLARE_PRIVATE(), Q_DECLARE_PUBLIC(), Q_Q(), and Q_D().
This makes it easier to inherit private classes (should I ever decide
to do that), and is more consistent with other Qt applications.

Removed the McRecoverQApplicationPrivate class instance. It wasn't
used for anything, so it was just wasting a pointer and CPU cycles.
Instead, the static functions in McRecoverQApplicationPrivate are
called directly using McRecoverQApplicationPrivate::function().