Added internal copies of libpng-1.6.8 and zlib-1.2.8.
zlib is not used directly by GCN MemCard Recover, so it's only
included if the internal libpng is also included.

TODO: Allow internal libpng with external zlib?

TODO: Add the APNG patch, and allow GCN MemCard Recover to dlopen()
libpng to be able to use an APNG-supporting libpng, even if it wasn't
present in the compile-time version.