[libgctools] APNG_dlopen.c: New module for dlopen()'ing APNG symbols.
GCN MemCard Recover now dlopen()'s the APNG symbols, so it will work
with libpng regardless of whether or not it supports APNG, and it will
take advantage of APNG support if it's present.

Note that the libpng major version must be identical in order to
prevent ABI issues.

TODO: Skip the dlopen() on Linux, since the symbol is probably
already loaded if it's present.