[mcrecover] MemCardModel: Fixed an obscure crash bug on close.
On my Windows XP VM, loading a memory card file with Japanese files and
then immediately  exiting resulted in a crash. I traced the crash to
MemCardModel::memCard_destroyed_slot(). It seems that the destroyed()
slot might be run *after* the MemCard is deleted, so it ends up
dereferencing an invalid pointer.

Cache the number of files in the MemCard in MemCardModelPrivate.
This eliminates the need to dereference the MemCard in the MemCard's
destroyed() slot.