[libgctools] GcImageWriter: Don't #include <setjmp.h>
libpng-1.2 will complain and #error if setjmp.h is included
before png.h is included.

This error was removed in libpng-1.4.0beta54, but since both
Debian and Ubuntu are stuck on libpng-1.2, we need to maintain
compatibility and hence have to remove the manual include.

Added "#ifdef PNG_SETJMP_SUPPORTED" around setjmp() blocks.