[gcbanner] New utility to extract the banner image from a GameCube opening.bnr file.
We're uploading GameCube banners to Sonic Retro, and it turns out that
some games use a different banner in opening.bnr compared to the banner
used in the save file.

This is a simple utility that, given the opening.bnr file, extracts the
RGB5A3 banner and converts it to PNG format. libgctools' GcImage and
GcImageWriter classes are used for this.