[gcbanner] gcbanner.cpp: Don't check for BNR1 or BNR2 sizes.
We check the magic number, which should be enough.
We also check the size read.

Wii save banner files have multiple different sizes, which can get
a bit annoying to check, so don't bother checking. (I'm thinking
the extra size is the icon, which may or may not be animated.)