[gcbanner] gcbanner.cpp: Use %u instead of %lu.
Printing size_t is problematic, since "long" on Win64
is 32-bit, not 64-bit. Additionally, glibc printf supports
%zu for size_t, but MSVCRT printf does not.

Cast size_t to unsigned int and avoid this problem.
I don't expect anyone to try running gcbanner on
files larger than 4 GB, and if they do, a bad printf
is the least of their problems.
not on Windows.