Fix split debug files in MSVC builds.
SplitDebugInformation.cmake: MSVC's linker automatically splits
debug information into a .pdb file, so there's no need to use
objcopy. In addition, objcopy is part of GNU binutils, and might
not be present on an MSVC system.

src/*/CMakeLists.txt: Install *.pdb instead of *.debug if this is
an MSVC build.