[mcrecover] MemCard: Reset the used block map when (re)loading the file list.
If the active block table is changed, the used block map will
probably be different. (This could happen with directory table
too, but is less likely.)

Make sure we update the used block map so the search function
will know what blocks are used and what blocks aren't if the
active directory or block table is changed.

resetUsedBlockMap(): New function, split out from the
MemCardPrivate constructor.