[mcrecover] MemCard, MemCardFile: Properly handle invalid reads.
If the card image is too small, handle an invalid read correctly
instead of just using whatever garbage is in memory. This prevents
random garbage from showing up as "files".

MemCard::loadSysInfo(): Use a 'for' loop to handle multiple tables.
This can probably be unrolled in optimized builds.

MemCardFile: Initialize gcBanner to nullptr to prevent an invalid
delete on MemCard close if the file is invalid. (Valid files that
don't have a banner end up setting gcBanner to nullptr, since
loadBannerImage() returns nullptr in that case.)