[mcrecover] MemCard::addLostFiles(): New function to add multiple lost files.
Instead of iterating through the QLinkedList<SearchData> in
McRecoverWindow, iterate through it in MemCard. This reduces the number
of emitted fileAboutToBeInserted() and filesInserted() signals, which
reduces the number of times QTreeView has to determine how wide to
make the columns, which massively improves performance.

Specifically, the lag that happened after scanning for lost files
on a card with lots of lost files is gone.