cmake/macros/Win32ImageVersionLinkerFlags.cmake: New macro for adding an image version to Win32 binaries.
The TARGET_VERSION property causes the version number to be appended to
the executable when built with MinGW, which isn't what we want.

This macro replaces the /version: (MSVC) or -Wl,etc (MinGW) flags in the
linker commands with the specified version numbers.

Updated mcrecover and gcbanner to use Win32ImageVersionLinkerFlags.

libgctools: Removed VERSION, since this is a static library.