ConvertTextFilesToNative.cmake: Disable if compiling with MSVC.
`unix2dos' isn't working correctly with MSVC. Also, if the user
is compiling with MSVC, the repository was probably checked out
with automatic CRLF conversions in the first place, so the files
don't need to be converted a second time.

src/CMakeLists.txt: Don't build `miniu2d' if compiling with MSVC,
since it won't be used.