[mcrecover] Removed FileComments.
Instead of a dedicated class, use a QString with a null character
separating GameDesc and FileDesc. This simplifies code in a few
places, and makes it possible to handle cards that don't have
separate game and file descriptions (or descriptions at all;
some older systems only have filenames.)

MemCardItemDelegate: If the string has an embedded null character,
treat it as GameDesc/FileDesc; otherwise, treat it as a regular
string and pass it to the default paint() function.

MemCardModel: Combine gameDesc() and fileDesc() into a single
string, with the two components separated by QChar(L'\0').

MemCardSortFilterProxyModel: TODO: Does the embedded null character
affect string sorting, or does it still work as expected? (That is,
does it stop at the '\0' or does it keep going?)