[mcrecover] card/Card.cpp: Initial Card base class.
MemCard will be rewritten as a subclass of Card and will
be renamed GcnCard. The idea is that the Card base class
will have basic low-level functions that can be used by
multiple implementations, and GcnCard will provide the
GameCube-specific stuff.

TODO: Also do this for MemCardFile. (Split into File [or
FileBase to prevent conflicts?] and GcnFile.)