GcnFile: Initialize gcBanner to nullptr.
This was lost in the transition to a File subclass, so if the file
didn't have a valid banner, gcBanner would be garbage data, resulting
in an invalid free on close.

This is a regression from commit 98282b3c862fbfd4b2efabc2299b3c281f8e4158.
(GcnFile: Rewrote as a subclass of File.)