File: Moved GcnFile::saveBanner(QString) here; added virtual functions for saveBanner(QIODevice*) and saveIcon().
The latter two functions require access to the raw GcImages, so the
GcImage stuff needs to be rewritten to be more generic.

McRecoverWindow: Removed GcnFile* casts, since they're no longer necessary.
(and in the case of saveBanner(QString), it doesn't work, since the function
isn't implemented in the subclass; also, it's not virtual.)