VmuFile: Initial ICONDATA_VMS loading.
Currently, only the monochrome icon is loaded.

DcImageLoader: Added a function to load monochrome icons as used
in ICONDATA_VMS.

TODO: Add an "icon" property to Card and use it in CardView.
Also add VMU card icon accessors to VmuFile so the VmuCard
can figure out which file has the card icon.