SAEditor: Initial Sonic Adventure file editor.
This will be the first editor for GCN MemCard Recover v0.3.

Currently, the SALevelStats' UI is implemented, and it shows
a grid for all 11 levels. The UI needs to be reduced to only
show levels that correspond to the selected character.

McRecoverWindow: Switch back to GCN files for now.

FileView: Temporarily hijack the "XML" button to load SAEditor.