EditorWindow: New editor window.
EditorWindow has a static function to create a new editor window
for a given GcnFile. This function creates an editor widget that
(will eventually) mark itself as compatible with a given game,
and then creates that widget.

SAEditor is now a widget, which is created when an SADX (GCN)
save is loaded.

FileView::on_btnEdit_clicked(): Create an EditorWindow for the
selected file.

File: Expose loadFileData() as a public function.

SAEditor::setFile(): Added a function to load a file.
Currently only supports SADX (GCN) files.