SAEditor: Load all save slots for DC files.
The save slot can't be selected yet, but this is a start
for multiple slot editing.

d->data[]: QVector containing allocated sa_save_file* data.
This is used for single-slot (GCN) and multi-slot (Dreamcast)
files. Note that since sa_save_file is a C struct, the data
is allocated using malloc() and free().

d->updateDisplay(): New function for updating the display.