SALevelClearCount: Initial "Level Clear Count" widget.
Based on SALevelStats, but without switchLevels(), since all levels
are present for all characters.

There's some UI stuff that needs to be cleaned up, but otherwise
it looks like it's usable. Also, the values need to be checked
against SASave.

SAEditor: Added an SALevelClearCount widget, and load the
Level Clear Count data in load().