SAEventFlagsView: Initial "Event Flags" view/editor.
It's called "View" because I'm using that to refer to an individual
component editor, whereas "Editor" is used for the whole thing,
e.g. "SAEditor".

SAEventFlags: Wrapper class for storing and manipulating event flags.
Includes a QHash<int, QString> to map event IDs to descriptions.
Most events are unknown, so it may be more efficient to use a QHash
instead of an array. (Then again, QHash requires extra lookups, so
maybe we should just load the strings into a temporary array...)

sa_event_flags.data.h: Array containing the actual ID to description
mapping of event flags. This will be changed to an auto-generated
file using the same format as SASave's Events.ini later.

SAEventFlagsModel: QAbstractListModel for displaying SAEventFlags.
This class isn't finished; items need to be checkable, and the
view has to be capable of being filtered by section. Each section
is a 64-event block. Thankfully, the sections are all contiguous,
so we don't need to use bitmasks.