SAEventFlagsView: Use load() and save() instead of eventFlags() and setEventFlags().
load() and save() take a pointer to the sa_save_slot instead of
the sa_event_flags. This matches the other SA1 editor widgets.

TODO: Eventually make this a "generic" events editor so it can
also be used for NPCs, in which case we'll have to change this
to take an arbitrary pointer to uint8_t and a size parameter.