SAEventFlagsView.ui: Set lstEventFlags.rootIsDecorated = false.
On Linux, when this is true (default), a tree structure is drawn
to the left of the checkboxes. We're just showing a flat list, so
the tree structure is unnecessary.

Strangely, on Windows 7 (with both Windows Classic and Aero themes),
the tree structure doesn't show up. I wonder why Qt decided not to
draw them on Windows...