[SonicAdventure] PageFilterModel: QSortFilterProxyModel for paging.
It's much easier to derive from QSortFilterProxyModel than from
QAbstractProxyModel, so that's what PageFilterModel is based on.

Two properties, currentPage and pageSize, are used to determine
what items should be visible. For event flags, pageSize is 64.

SAEventFlagsView: Added a QTabBar for switching pages. Note that
Qt Designer doesn't have a QTabBar widget, so most of the QTabBar
initialization is in SAEventFlagsView's constructor.