[SonicAdventure] PageFilterModel: Added currentPageChanged signal and pageCount property.
pageCount is read-only and based on the model; thus, it isn't stored.
It's basically rows/pageSize, plus one if the last page isn't full.
(I used the little-known div() function to get both the quotient and
the remainder at the same time, which is a feature of the division
instruction on most CPUs.)

The currentPageChanged() signal is useful because it allows us to ensure
the tabBar's selected tab is adjusted if we adjust the filter model's
current page index directly.

currentPageChanged() is also emitted if the pageSize is changed to a
value that causes the current page to disappear. For example, if the
pageSize is 10 and currentPage is 9, setting pageSize to 20 will
force currentPage to be 4.

calcPageOffsets(): Fixed up some issues. Only execute the first block
if sourceRowCount == 0, since we want to adjust the current page if it's
out of range.

SAEventFlagsView: Connect pageFilterModel->currentPageChanged() to
d->ui.tabBar->setCurrentIndex().