BitFlags::setAllFlags(): Use the bit count internally.
We were always using d->flags.count() in the for loop, which could cause
a segfault if data wasn't large enough.

Additionally, flagsChanged() was emitting the byte count, so only the
first 1/8th of flags were updated in the model. The same is true of
the return value, which now correctly returns the number of flags
updated instead of the number of bytes.