[SonicAdventure] BitFlagsView: Expose the pageSize() property.
Allow SAEditor to set BitFlagsView's pageSize. Since the pageFilterModel
is private to BitFlagsView, we have to expose the property separately.

PageFilterModel:
- Allow a pageSize of 0. This effectively makes the PageFilterModel
  a no-op, since it has a single page containing all of the rows.
- Move the default pageSize to a private static const variable,
  PageFilterModel::defualtPageSize.

SAEditor: Set saEventFlagsView's pageSize to 64, and saNPCFlagsView's
pageSize to 0. None of the NPC flags are known, so they should be
shown on a single page.

TODO: Show NPC flag IDs.