Card: Cleaned up the "encoding" field.
Instead of requesting a QTextCodec for the specified encoding, always
try decoding as Shift-JIS, then fall back to cp1252 if it fails. This
is necessary for VMUs, since the file system doesn't specify the filename
encoding. It also helps with GameCube, since we can't rely on either the
card's encoding or the file's region to determine the actual encoding.

Added Encoding::ENCODING_UNKNOWN for "unknown" encoding. This should
be used for VMUs, since there's no card-wide encoding.

Moved VmuFilePrivate::decodeText_SJISorCP1252() to FilePrivate.

TODO:
- CardView: Hide the "Encoding:" label for ENCODING_UNKNOWN.
- McRecoverWindow: Need to add a "hasJapaneseCharacters" function.