[SonicAdventure] Added initial "save" functionality to SALevelStats and SALevelClearCount.
This currently means saving to the internal variables. Since Card and File
don't have write support yet, the actual file cannot be saved.

For now, this means the data will be retained when switching between
file slots on Dreamcast files, and when switching between characters
in SALevelStats.

TODO: SALevelStats: Add signals for the widgets and remove the need
to call d->saveCurrentStats() from SALevelStats::save().