[SonicAdventure] SAAdventure::load(): Initial load() implementation.
Currently only loads the life counters.

initCharacters(): Don't add a life counter for the "Unused" character,
since it isn't stored in the save file.

Rearranged the rows to match a more logical order, with "Super Sonic"
and "Unused" at the end. Added some maps to map the row indexes to
the character indexes, and then to the save file indexes.

TODO: The save file indexes match the row indexes, so this map
might not be necessary.