[SonicAdventure] SAGeneral: Added a field for "Black Market Rings".
This field is hidden by default, but it's shown if an SADX extra save
is loaded, similar to Metal Sonic in SALevelStats.

Added loadDX() and saveDX() functions for the SADX extra save data.

SAEditor.ui: Removed the "General" tab in "SADX Extras".
In SASave, this only contained the "Black Market Rings" field.
Since this is now in SAGeneral, there's no need for the extra
tab in "SADX Extras".