[SonicAdventure] SAEditor: Removed the "SADX Extras" tab.
Metal Sonic's data is editable (or will be editable) in the other tabs.
Black Market rings is in the General tab.

The only thing left is Missions, which can be its own tab instead of
SADX Extras.

Note that we're now using addTab() to add the tab to the end of the
QTabWidget. Using insertTab() doesn't work properly for this, and it
results in weird drawing issues when switching back and forth between
files with SADX extras and files without SADX extras.