[SonicAdventure] EditorWidget: New base class for editor widgets.
The EditorWidget base class will handle save slot management.
In addition to the regular save slots, it will also provide an option
for a "general" settings tab.

SAEditor: Rebased this class as a subclass of EditorWidget.

TODO: Add a toolbar to EditorWindow that will have buttons for each
save slot, effectively replacing SlotSelector.