[SonicAdventure] SALevelClearCount: Moved the QScrollArea into the widget.
This allows us to have the header remain in place while scrolling the
rest of the widgets.

Set the alignment of the header labels in Qt Designer rather than
calling setAlignment() in the SALevelClearCount constructor.

TODO: The header labels aren't aligned correctly if the vertical scrollbar
is visible. Extra spacing needs to be added in that case.