[SonicAdventure] SALevelClearCount: Make sure the widgets stay at the top of the scroll area.
Added a QSpacerItem at the bottom to ensure that if the window is
expanded past the total size of the QScrollArea, the widgets stay
at the top instead of adding space between each row.

New #define TOTAL_CHARACTERS. Use this macro throughout the file
wherever the total number of characters is needed.

SALevelClearCount.ui: Mark scrlLevelsContent as H=Expanding, V=Minimum.
This didn't help with spacing (the QSpacerItem was still needed), but
it should still be set.