[mcrecover] File: Use system-independent mode bits.
These map to GCN mode bits (with a small shift).
VMU mode bits require a bit of fiddling, but it works.

The EXEC bit is currently only used for VMU.

The GLOBAL bit is GCN-specific, though I'm not sure if it's even
an actual bit, since I've never seen it in use.