[libgctools] gif_writeARGB32Image(): New function to write ARGB32 images to GIF.
This uses giflib's GifQuantizeBuffer() function.

When using giflib-4.2 or earlier, we compile a copy of giflib-5.1.2's
quantize.c (with some changes) and use that instead. This is needed
because giflib-4.2 removed quantize.c. Earlier versions have it, but
in order to maintain some sanity (4.1 vs. 4.2 issues), we're going to
use the internal version on all giflib older than 5.x.