[libgctools] GIF_dlopen.c: Implemented GifDlGetUserData().
When USE_INTERNAL_GIF, this is implemented as a macro that directly
accesses GifFileType::UserData.

When !USE_INTERNAL_GIF, this checks the version of giflib in use,
casts the GifFileType* to the appropriate version struct, and
retrieves the UserData pointer.

GIF_dlopen.struct.h: Added a forward declaration for struct SavedImage.
This fixes the !USE_INTERNAL_GIF build (to a point; linking fails due to
missing wrapper functions).

GcImageWriter_GIF.cpp: Use GifDlGetUserData().