[libgctools] GIF_dlopen.c: Use .so version numbers internally instead of 51, 50, etc.
This improves the chances for switch/case to be optimized, and it makes
a few things easier to deal with. For example, we don't have to map the
"real" version number to the .so version when loading giflib.

GIF_dlopen.struct.h: New enum GifLib_Version_t, which has enum constants
for the various .so versions. Note that .so version 5 doesn't actually
exist, since both giflib-4.1 and giflib-4.2 are .so version 4. The two
versions can be distinguished by the presence of additional error handling
functions in giflib-4.2.