[mcrecover] Win7TaskbarList: New TaskbarButtonManager for Windows 7+.
This uses the ITaskbarList3 interface.

McRecoverQApplication_win32.cpp: Register the "TaskbarButtonCreated"
message. This message is sent by Windows once a window's taskbar
button is created to indicate that we can instantiate the corresponding
ITaskbarList3 object.

McRecoverWindow: We need to wait for the "TaskbarButtonCreated" message
before creating the TaskbarButtonManager, so listen for the message in
the winEvent() filter.

FIXME: winEvent() has been replaced by nativeEvent() in Qt 5.