[mcrecover] VmuCard: Added the default Dreamcast BIOS icons.
Card: Added a function icon() to get the card's icon. Currently,
this is only implemented for VmuCard, and only for default icons.
(TODO: Add support for "ICONDATA_VMS".)

CardView: Added card icon and type display. The type display is
currently not updated, since the productName() field is too long
for the display. I'll need to add a "short" product name field later.