[SonicAdventure] SAEditWidget: Added modification properties and signals.
Note that signals can't be emitted from const functions, and since we're
emitting a notification signal when 'modified' changes, not just when
'modified' is set, we have to remove 'const' from the save functions.

TODO: Do we really need to emit notification signals when 'modified' is
cleared? (I think we should, since it lets us keep the NOTIFY signal
in Q_PROPERTY.)