[edit] BitFlags: Added support for translations.
If tr_ctx is specified, messages will be translated using the specified
translation context. Otherwise, they'll be converted from Latin-1. (Note
that this conversion may be slower and/or use more memory than previously,
since there's no implicit string sharing.)

SADataEvents.c: Added QT_TRANSLATE_NOOP() using "SADataEvents" as the
context. Note that this context is different from SAEventFlags, which
uses the context "SAEventFlags". (The latter is the view, not the data
source.)

Updated SAEventFlags and SANPCFlags to handle this change.