[mcrecover] mcrecover.cpp: Convert the filename from native separators.
This fixes an issue on Windows where the full pathname was displayed
as the basename, since the pathname contained backslashes instead of
slashes.

Use QCoreApplication::arguments() instead of argv[]. On Windows, Qt
creates this QStringList using GetCommandLineW(), so it supports
Unicode, whereas argv[] does not.

Note that McRecoverQApplication_win32.cpp's argv[] code is still
needed in order to provide argc and argv[], even though it's not
actually used for argument processing. (I think qWinMain() might
also be what initializes QCoreApplication::arguments(), but I'm
not entirely sure.)