[mcrecover] Card: Preliminary write support.
This merely adds an option to make the card writable, which reopens
the file with write permission.

File: Add an isReadOnly() function that returns whether the underlying
card is read-only.

McRecoverWindow: Added an "Allow Write" checkbox. This checkbox is
disabled if no card is loaded or if the card has errors.

TODO: Currently, a second QFile is opened with write permission, then
the original QFile is closed. Does this work on Windows?