[edit] BitFlags::allFlags(): Fix incorrectly reversed bitfields.
We need to shift bits in from the left side (0x80), not the right side
(0x01). Otherwise, the bit flags will be reversed, resulting in incorrect
data every other time the bit flags are loaded and saved.

This fixes the 128-byte area at 0x16E0 in SADX GCI files being incorrect.
(This is the event flags and NPC flags area.)