[mcrecover] McRecoverWindow: Use selectionChanged() instead of currentRowChanged().
This fixes the issue where specifying a filename on the command line
results in the first file on the card being selected even though it
really isn't selected.

FIXME: QItemSelection::indexes() *crashes* in MSVC debug builds once
the function returns. Currently using a hackish method to get the
selected item, which is somewhat off if the user unselects an item
while multiple items are selected. This is basically the same behavior
as before, and it works better overall.