[mcrecover] TableSelect: Dynamically generate widgets based on DAT/BAT count.
This will allow us to support devices with any number of directory
and/or block tables.

TODO: Switch from GcnCard* to Card*, since we can support a single
table now. (Not that switching tables is useful if only a single
table is present, but we can at least show the table status.)