Removed GcnDateTime.
It might have been useful at first when the program could only read
GCN memory card files, but the addition of VMU files made it a bit
more complicated.

Switch to QDateTime everywhere and add a few conversion functions to
convert to and from GCN timestamps.

For VMU, there's currently only a function to convert from VMU to
QDateTime at the moment.