[mcrecover] McRecoverWindow: Set theme icons in the designer file.
This was added in Qt 4.8. Since we're using Qt 5 now, we can be sure
that this feature always exists, so use it instead of setting Oxygen
icons and then changing them to the theme icons.

This should reduce application load time by a bit.

NOTE: Qt Designer is adding and/or changing mnemonics when editing
McRecoverWindow for some reason. It's also removing extracomments
from QAction shortcut keys.