[mcrecover] VarReplace::ApplyModifiers(): Set the QDateTime to UTC.
I know it worked before without this, but when testing the timestamp
modifiers with some new definitions, I noticed that it had an incorrect
time. It was off by exactly my current timezone offset.

This appears to fix it, and it doesn't cause problems with file
definitions that don't have timestamp modifiers.